APDL: Prog HP pass 1
remove some casts from when I was doing float stuff, consistently retype HP
reorder one of the hook calls, originally intended for latency

+++ APDeathLink.h
@@ -25,7 +25,7 @@ namespace APDeathLink
 	void recvHP();
 	void prog_hp_update();
 	void prog_hp_reset();
-	void setHP(uint8_t HP);
+	void setHP(int HP);
 
 	void ImGuiTab();
 };
